+++ bugsnag/delivery.py
@@ -28,8 +28,8 @@ except ImportError:
 
 DEFAULT_ENDPOINT = 'https://notify.bugsnag.com'
 DEFAULT_SESSIONS_ENDPOINT = 'https://sessions.bugsnag.com'
-HUB_ENDPOINT = 'https://notify.insighthub.smartbear.com'
-HUB_SESSIONS_ENDPOINT = 'https://sessions.insighthub.smartbear.com'
+SECONDARY_ENDPOINT = 'https://notify.bugsnag.smartbear.com'
+SECONDARY_SESSIONS_ENDPOINT = 'https://sessions.bugsnag.smartbear.com'
 
 __all__ = ('default_headers', 'Delivery')
 
@@ -86,8 +86,8 @@ class Delivery:
         """
         if ((config.endpoint != DEFAULT_ENDPOINT and
              config.session_endpoint == DEFAULT_SESSIONS_ENDPOINT) or
-            (config.endpoint != HUB_ENDPOINT and
-             config.session_endpoint == HUB_SESSIONS_ENDPOINT)):
+            (config.endpoint != SECONDARY_ENDPOINT and
+             config.session_endpoint == SECONDARY_SESSIONS_ENDPOINT)):
             if not self.sent_session_warning:
                 warnings.warn('The session endpoint has not been configured. '
                               'No sessions will be sent to Bugsnag.')